refactor: 자동차 리스트를 관리하기 위한 클래스 추가
