+++ src/main/java/racingcar/model/Cars.java
@@ -1,6 +1,5 @@
 package racingcar.model;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import racingcar.constant.ErrorMessages;
@@ -8,44 +7,18 @@ import racingcar.util.DefaultRandomGenerator;
 import racingcar.util.RandomGenerator;
 
 public class Cars {
-    private static final String DELIMITER_COMMA = ",";
-    private static final int MIN_LIST_SIZE = 2;
     private static final int MOVE_THRESHOLD = 4;
 
     private final List<Car> carList;
     private final RandomGenerator randomGenerator;
 
-    public Cars(String carNames) {
-        this(carNames, new DefaultRandomGenerator());
+    public Cars(List<Car> carList) {
+        this(carList, new DefaultRandomGenerator());
     }
 
-    public Cars(String carNames, RandomGenerator randomGenerator) {
-        List<Car> tempCarList = convertCarNamesToList(carNames).stream()
-                .map(Car::new)
-                .toList();
-        validateListSize(tempCarList);
-        validateNoDuplicateNames(tempCarList);
-        this.carList = tempCarList;
+    public Cars(List<Car> carList, RandomGenerator randomGenerator) {
+        this.carList = carList;
         this.randomGenerator = randomGenerator;
-    }
-
-    private List<String> convertCarNamesToList(String carNames) {
-        return Arrays.asList(carNames.split(DELIMITER_COMMA));
-    }
-
-    private void validateListSize(List<Car> carList) {
-        if (carList.size() < MIN_LIST_SIZE) {
-            throw new IllegalArgumentException(ErrorMessages.INVALID_NAME_LIST_SIZE);
-        }
-    }
-
-    private void validateNoDuplicateNames(List<Car> carList) {
-        long distinctCount = carList.stream()
-                .distinct()
-                .count();
-        if (carList.size() != distinctCount) {
-            throw new IllegalArgumentException(ErrorMessages.DUPLICATE_NAME);
-        }
     }
 
     public void tryMoveAll() {
